kernel: Boost to the max for a short amount of time when zygote forks
In order to boost during the period from when a new app is forked to the
time that it is added to the top-app cpuset (which is a signal to
subsequently boost again), boost for a short amount of time when zygote
forks. Not all processes forked from zygote are top apps, hence the
decoupled boosting setup.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[0ctobot: Adapted for msm-4.14]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: LibXZR <xzr467706992@163.com>
Signed-off-by: dlwlrma123 <alexfinhart@gmail.com>