AnyKernel3: Mafia Final
I want to change localversion to chaeyoung for final but it such a drag to edit it again especially when im using github website, so anyway this is really the last.. if you're a TG user and probably reading this, CY kernel is already EOL(end-of-life) so please stop asking for support. The changes (made things stable) i made are for users who wants to use CY even if isn't updated anymore. Thank you and have a nice day.